refactoring the view models to move the enumeration classes back to the core musical library so both UI's have the same classes
